requirements.txt
Flask>=3.0.0
selenium>=4.0.0
python-dotenv>=1.0.0
pandas>=2.2.0
openpyxl>=3.1.0
xlrd>=2.0.1
